fix: move string manipulation methods from libcommonserver to libcommon
# Conflicts:
#	src/libsyncengine/update_detection/file_system_observer/computefsoperationworker.cpp
