+++ src/gui/localfolderdialog.cpp
@@ -25,6 +25,7 @@
 #include "guirequests.h"
 #include "libcommon/utility/utility.h"
 #include "libcommongui/matomoclient.h"
+#include "libcommonserver/utility/utility.h"
 
 #include <QBoxLayout>
 #include <QDesktopServices>
@@ -203,9 +204,9 @@ void LocalFolderDialog::updateUI() {
 
         if (virtualFileMode == VirtualFileMode::Win || virtualFileMode == VirtualFileMode::Mac) {
             // Check file system
-            QString fsName(KDC::CommonUtility::fileSystemName(_localFolderPath));
-            _folderCompatibleWithLiteSync = ((virtualFileMode == VirtualFileMode::Win && fsName == "NTFS") ||
-                                             (virtualFileMode == VirtualFileMode::Mac && fsName == "apfs"));
+            _folderCompatibleWithLiteSync =
+                    ((virtualFileMode == VirtualFileMode::Win && CommonUtility::isNTFS(QStr2Path(_localFolderPath))) ||
+                     (virtualFileMode == VirtualFileMode::Mac && CommonUtility::isAPFS(QStr2Path(_localFolderPath))));
             if (!_folderCompatibleWithLiteSync) {
                 _warningLabel->setText(tr(R"(This folder is not compatible with Lite Sync.<br>
 Please select another folder. If you continue Lite Sync will be disabled.<br>
