+++ src/libcommon/utility/utility.cpp
@@ -24,8 +24,14 @@
 #include <system_error>
 #include <sys/types.h>
 
-#ifdef Q_OS_UNIX
+#if defined(KD_MACOS)
 #include <sys/statvfs.h>
+#include <sys/mount.h>
+#elif defined(KD_LINUX)
+#include <sys/statvfs.h>
+#include <sys/statfs.h>
+#elif defined(KD_WINDOWS)
+#include <fileapi.h>
 #endif
 
 
@@ -184,17 +190,96 @@ const std::string &CommonUtility::currentVersion() {
     return str;
 }
 
-QString CommonUtility::fileSystemName(const QString &dirPath) {
-    if (QDir dir(dirPath); dir.exists()) {
-        if (const QStorageInfo info(dirPath); info.isValid()) {
-            return info.fileSystemType();
+static std::unordered_map<std::string, std::string> rootFsTypeMap;
+std::string getRootFsType(const SyncPath &targetPath) {
+    auto it = rootFsTypeMap.find(targetPath.root_name().string());
+    if (it == rootFsTypeMap.end()) {
+        const std::string fsType = CommonUtility::fileSystemName(targetPath);
+        const auto [it2, inserted] = rootFsTypeMap.try_emplace(targetPath.root_name().string(), CommonUtility::toUpper(fsType));
+        if (!inserted) {
+            return {};
         }
-    } else {
-        dir.cdUp();
-        return fileSystemName(dir.path());
+        it = it2;
     }
+    return it->second;
+}
 
-    return {};
+bool CommonUtility::isNTFS(const SyncPath &targetPath) {
+    static const std::string ntfs("NTFS");
+    return getRootFsType(targetPath) == ntfs;
+}
+
+bool CommonUtility::isAPFS(const SyncPath &targetPath) {
+    static const std::string apfs("APFS");
+    return getRootFsType(targetPath) == apfs;
+}
+
+bool CommonUtility::isFAT(const SyncPath &targetPath) {
+    static const std::string fat("FAT");
+    return contains(getRootFsType(targetPath), fat);
+}
+
+std::string CommonUtility::fileSystemName(const SyncPath &targetPath) {
+#if defined(KD_MACOS)
+    struct statfs stat;
+    if (statfs(targetPath.root_path().native().c_str(), &stat) == 0) {
+        return stat.f_fstypename;
+    }
+#elif defined(KD_WINDOWS)
+    TCHAR szFileSystemName[MAX_PATH + 1];
+    DWORD dwMaxFileNameLength = 0;
+    DWORD dwFileSystemFlags = 0;
+
+    if (GetVolumeInformation(targetPath.root_path().c_str(), NULL, 0, NULL, &dwMaxFileNameLength, &dwFileSystemFlags,
+                             szFileSystemName, sizeof(szFileSystemName)) == TRUE) {
+        return ws2s(szFileSystemName);
+    } else {
+        // Not all the requested information is retrieved
+        DWORD dwError = GetLastError();
+        LOGW_WARN(logger(), L"Error in GetVolumeInformation for " << formatSyncName(targetPath.root_name()) << L" ("
+                                                                  << utility_base::getErrorMessage(dwError) << L")");
+
+        // !!! File system name can be OK or not !!!
+        return ws2s(szFileSystemName);
+    }
+#elif defined(KD_LINUX)
+    struct statfs stat;
+    if (statfs(targetPath.root_path().native().c_str(), &stat) == 0) {
+        const auto formatFsName = [](const std::string &prettyName, long fsCode) {
+            std::stringstream stream;
+            stream << std::hex << fsCode;
+            return prettyName + " | 0x" + stream.str();
+        };
+        switch (stat.f_type) {
+            case 0x137d:
+                return formatFsName("EXT(1)", stat.f_type);
+            case 0xef51:
+                return formatFsName("EXT2", stat.f_type);
+            case 0xef53:
+                return formatFsName("EXT2/3/4", stat.f_type);
+            case 0xbad1dea:
+            case 0xa501fcf5:
+            case 0x58465342:
+                return formatFsName("XFS", stat.f_type);
+            case 0x9123683e:
+            case 0x73727279:
+                return formatFsName("BTRFS", stat.f_type);
+            case 0xf15f:
+                return formatFsName("ECRYPTFS", stat.f_type);
+            case 0x4244:
+                return formatFsName("HFS", stat.f_type);
+            case 0x5346544e:
+                return formatFsName("NTFS", stat.f_type);
+            case 0x858458f6:
+                return formatFsName("RAMFS", stat.f_type);
+            default:
+                return formatFsName("Unknown-see corresponding entry at https://man7.org/linux/man-pages/man2/statfs.2.html",
+                                    stat.f_type);
+        }
+    }
+#endif
+
+    return "Error";
 }
 
 void CommonUtility::resetTranslations() {
@@ -211,6 +296,81 @@ void CommonUtility::resetTranslations() {
             _qtTranslator = nullptr;
         }
     }
+}
+
+bool CommonUtility::startsWith(const std::string &str, const std::string &prefix) {
+    return str.size() >= prefix.size() &&
+           std::equal(prefix.begin(), prefix.end(), str.begin(), [](char c1, char c2) { return c1 == c2; });
+}
+
+bool CommonUtility::startsWithInsensitive(const std::string &str, const std::string &prefix) {
+    return str.size() >= prefix.size() && std::equal(prefix.begin(), prefix.end(), str.begin(), [](char c1, char c2) {
+               return std::tolower(c1, std::locale()) == std::tolower(c2, std::locale());
+           });
+}
+
+bool CommonUtility::endsWith(const std::string &str, const std::string &suffix) {
+    return str.size() >= suffix.size() && std::equal(str.begin() + static_cast<long>(str.length() - suffix.length()), str.end(),
+                                                     suffix.begin(), [](char c1, char c2) { return c1 == c2; });
+}
+
+bool CommonUtility::endsWithInsensitive(const std::string &str, const std::string &suffix) {
+    return str.size() >= suffix.size() &&
+           std::equal(str.begin() + static_cast<long>(str.length() - suffix.length()), str.end(), suffix.begin(),
+                      [](char c1, char c2) { return std::tolower(c1, std::locale()) == std::tolower(c2, std::locale()); });
+}
+
+bool CommonUtility::contains(const std::string &str, const std::string &substr) {
+    return str.find(substr) != std::string::npos;
+}
+
+std::string CommonUtility::toUpper(const std::string &str) {
+    std::string upperStr(str);
+    // std::ranges::transform(str, upperStr.begin(), [](unsigned char c) { return std::toupper(c); });   // Needs gcc-11
+    std::transform(str.begin(), str.end(), upperStr.begin(), [](unsigned char c) { return std::toupper(c); });
+    return upperStr;
+}
+
+#if defined(KD_WINDOWS)
+bool Utility::startsWithInsensitive(const SyncName &str, const SyncName &prefix) {
+    return str.size() >= prefix.size() && std::equal(prefix.begin(), prefix.end(), str.begin(), [](SyncChar c1, SyncChar c2) {
+               return std::tolower(c1, std::locale()) == std::tolower(c2, std::locale());
+           });
+}
+
+bool Utility::startsWith(const SyncName &str, const SyncName &prefix) {
+    return str.size() >= prefix.size() &&
+           std::equal(prefix.begin(), prefix.end(), str.begin(), [](SyncChar c1, SyncChar c2) { return c1 == c2; });
+}
+
+bool Utility::endsWith(const SyncName &str, const SyncName &suffix) {
+    return str.size() >= suffix.size() && std::equal(str.begin() + str.length() - suffix.length(), str.end(), suffix.begin(),
+                                                     [](char c1, char c2) { return c1 == c2; });
+}
+
+bool Utility::endsWithInsensitive(const SyncName &str, const SyncName &suffix) {
+    return str.size() >= suffix.size() &&
+           std::equal(str.begin() + str.length() - suffix.length(), str.end(), suffix.begin(),
+                      [](char c1, char c2) { return std::tolower(c1, std::locale()) == std::tolower(c2, std::locale()); });
+}
+#endif
+
+bool CommonUtility::isDescendantOrEqual(const SyncPath &potentialDescendant, const SyncPath &path) {
+    if (path == potentialDescendant) return true;
+    for (auto it = potentialDescendant.begin(), it2 = path.begin(); it != potentialDescendant.end(); ++it, ++it2) {
+        if (it2 == path.end()) {
+            return true;
+        }
+        if (*it != *it2) {
+            return false;
+        }
+    }
+    return false;
+}
+
+bool CommonUtility::isStrictDescendant(const SyncPath &potentialDescendant, const SyncPath &path) {
+    if (path == potentialDescendant) return false;
+    return isDescendantOrEqual(potentialDescendant, path);
 }
 
 QString CommonUtility::getIconPath(const IconType iconType) {
