+++ src/libsyncengine/syncpal/tmpblacklistmanager.cpp
@@ -78,7 +78,7 @@ void TmpBlacklistManager::blacklistItem(const NodeId &nodeId, const SyncPath &re
 
     std::list<NodeId> toBeRemoved;
     for (const auto &[id, errorInfo]: errors) {
-        if (Utility::isDescendantOrEqual(errorInfo.path, relativePath) && id != nodeId) {
+        if (CommonUtility::isDescendantOrEqual(errorInfo.path, relativePath) && id != nodeId) {
             toBeRemoved.push_back(id);
         }
     }
@@ -120,14 +120,14 @@ void TmpBlacklistManager::removeItemFromTmpBlacklist(const SyncPath &relativePat
 
     // Find the node id of the item to be removed
     for (const auto &[nodeId, tmpInfo]: _localErrors) {
-        if (Utility::isDescendantOrEqual(tmpInfo.path, relativePath)) {
+        if (CommonUtility::isDescendantOrEqual(tmpInfo.path, relativePath)) {
             localId = nodeId;
             break;
         }
     }
 
     for (const auto &[nodeId, tmpInfo]: _remoteErrors) {
-        if (Utility::isDescendantOrEqual(tmpInfo.path, relativePath)) {
+        if (CommonUtility::isDescendantOrEqual(tmpInfo.path, relativePath)) {
             remotedId = nodeId;
             break;
         }
@@ -150,7 +150,7 @@ void TmpBlacklistManager::removeItemFromTmpBlacklist(const NodeId &nodeId, const
 
 bool TmpBlacklistManager::isTmpBlacklisted(const SyncPath &path, const ReplicaSide side) const {
     for (auto &errors = side == ReplicaSide::Local ? _localErrors : _remoteErrors; const auto &errorInfo: errors) {
-        if (Utility::isDescendantOrEqual(path, errorInfo.second.path)) return true;
+        if (CommonUtility::isDescendantOrEqual(path, errorInfo.second.path)) return true;
     }
 
     return false;
